render doc push

+++ VirtualWorlds/ModuleProject/Light.h
@@ -4,6 +4,6 @@
 struct DirectionalLight {
 	glm::vec3 direction = glm::vec3(0.0f);
 	glm::vec3 colourIntensities;
-	int ambient;
-
+	float ambient;
 };
+
